fix detection of openstack project
we push as "public" visibility for production images.
currently, this is decided based on whether you are authenticated
as the openstack project. Detecting it directly from the `connection`
object does not work for app credentials. Instead, get the current
authenticated project ID, then look it up to see if the name matches.

This is still somewhat brittle, and should be made more explicit in the
invocation.