allow builds without authentication to tacc
to avoid duplicate work, artifacts.py attempts to list already
existing images from tagg. However, a user may not care about this,
and it introduces a dependency on authentication and access to tacc.

This change just sets the value to Null if the connection fails, but future
changes should depend only on public images, or instead use a dedicated cache
configuration of some kind.