build variant of ubuntu24.04 with hwe kernel (#14)
build variant of ubuntu24.04 with hwe 6.11 kernel

We dicovered that the 6.8 kernels that ship with ubuntu24.04
(at least 6.8.0-47 - 6.8.0-53), have a panic with the amdgpu kernel
module when an AMD MI100 gpu is present on the system. The panic is
shown in the dmesg logs from boot up, but causes an infinite hang
on shutdown/reboot, due as udev also fails.

This bug appears to be fixed in 6.10, see:
https://lore.kernel.org/dri-devel/20240413213708.3427038-1-alexander.deucher@amd.com/

and we can get 6.11 by installing ubuntu's hwe kernel.